Move ScenarioModelBuilder.kt to model package
